src/requirements.txt
django                          
django-widget-tweaks                                 
Pillow
djangoajax
django-star-ratings
#sendgrid
django-tinymce
python-decouple

    
# https://infinite-scroll.com   
# auth - token ++
# api - call +
# serializer +++
# docker       
